third_party/blink/web_tests/external/wpt/css/css-mixins/contents-nested-declarations-expected.txt

This is a testharness.js-based test.
[FAIL] Can mix declarations into pseudo-elements via @contents
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
[FAIL] Nested declarations from @contents have top-level specificity behavior
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
